Added two new letter codes that will now have an effect if present in the comma-separated Type field of a mod's ini file:   - The 'M' type had no meaning before this commit.  It now represents 'Manual download required'.   - The 'W' type had no meaning before this commit.  It now represents 'Warning'.
Added 'or warnings' to the expert mods pop-up confirmation window when you click 'Continue' to start a new installation if you have mods selected that either have only 'E' as their Type or have a 'W' somewhere in their Type list.  Translations for EN, GE and RU are all updated with this addition.

Added background highlighting for mod names in the "choose your components" window, if the mod has either 'W' or 'M' in its Type:
  - The 'W' type gets a yellow background color.
  - The 'M' type gets a light grey background color.

If a mod has both 'W' and 'M' in its Type, only the 'W' background color will be used.

Also fixed some typos in the au3 files, just because.